Shortened the warning messages to avoid a bug in the testing framework.
A line containing an expected-warning annoation
in the VariableNames.cpp test made the test fail because it was
too long, which seems to trigger a bug in the testing
framework that let the test fail.
The current fix is to reduce the the length of the warning message
until this is triaged and fixed.